delete code related old logic
