+++ DeliveryConstants.py
@@ -1,6 +1,5 @@
 emailGroups = {
     "standard": "igo",  # will be added to every delivery, together with project LabHead and PI
-    "raw": "",  # only standard list
     "impact": "pipelineSupport",
     "access": "access",
     "wesWithCCS": "ccs",
@@ -17,6 +16,11 @@ emailAddresses = {
     "bic": "bic-request@cbio.mskcc.org",
     "ccs": "zzPDL_CMO_TEMPO_Support@mskcc.org,skicmopm@mskcc.org",
 }
+
+DEFAULT_ADDRESS = "zzPDL_SKI_IGO_DATA@mskcc.org"
+SKI_SENDER_ADDRESS = "igoski@mskcc.org"
+MSKCC_ADDRESS = "YOUR_MSKCC_ADDRESS"
+NO_PM = "NO PM"
 
 genericSubject= "[IGO Data] %s Project %s "
 
@@ -36,43 +40,27 @@ genericAnalysisContent = "Hello,<br><br>IGO processing is now complete. FASTQ fi
 
 crisprAddon = "<br>In addition, you will receive the results of the CRISPResso analysis, generated by the CRISPResso2 software. <br> For a detailed description of the output please consult https://crispresso.pinellolab.partners.org/help "
 
-DEFAULT_ADDRESS = "naborsd@mskcc.org,mcmanamd@mskcc.org"
-NEW_DEFAULT_ADDRESS = "zzPDL_SKI_IGO_DATA@mskcc.org"
-SENDER_ADDRESS = "naborsd@mskcc.org"
-SKI_SENDER_ADDRESS = "igoski@mskcc.org"
-MSKCC_ADDRESS = "YOUR_MSKCC_ADDRESS"
-NO_PM = "NO PM"
-REDELIVERY_MESSAGE = """Hello,
+FOOTER = "<br><br>Thank you, <br><a href='http://genomics.mskcc.org/'>Integrated Genomics Operation</a><br><a href='https://www.mskcc.org'>Memorial Sloan Kettering Cancer Center</a><br>Follow us on <a href='https://www.instagram.com/genomics212/?hl=en'>Instagram</a> and <a href='https://twitter.com/genomics212?lang=en'>Twitter</a>!<br><br>Please rate your delivery experience <a href='https://genomics.mskcc.org/feedback/data-delivery'>here</a><br>"
 
-We have linked additional fastqs to this project. Often this is caused because we resequenced a pool and already sufficiently-sequenced samples received additional reads. There are no additional charges for these reads.
-
-"""
-FOOTER = """Integrated Genomics Operations,
-Center for Molecular Oncology,
-Memorial Sloan Kettering Cancer Center, New York.
-http://cmo.mskcc.org/"""
-
-def getAliases(self):
-        aliases = defaultdict(str)
-        aliases["a-haimovitz-friedman@ski.mskcc.org"] = "haimovia"
-        aliases["a-zelenetz@mskcc.org"] = "zeleneta"
-        aliases["a-zelenetz@ski.mskcc.org"] = "zeleneta"
-        aliases["d-scheinberg@ski.mskcc.org"] = "scheinbd"
-        aliases["f-giancotti@ski.mskcc.org"] = "giancotf"
-        aliases["j-massague@ski.mskcc.org"] = "massaguej"
-        aliases["k-anderson@ski.mskcc.org"] = "kanderson"
-        aliases["m-baylies@ski.mskcc.org"] = "bayliesm"
-        aliases["m-jasin@ski.mskcc.org"] = "jasinm"
-        aliases["m-mcdevitt@ski.mskcc.org"] = "m-mcdevitt"
-        aliases["m-moore@ski.mskcc.org"] = "moorem"
-        aliases["m-ptashne@ski.mskcc.org"] = "ptashne"
-        aliases["m-sadelain@ski.mskcc.org"] = "sadelaim"
-        aliases["m-van-den-brink@ski.mskcc.org"] = "vandenbm"
-        aliases["p-tempst@ski.mskcc.org"] = "tempstp"
-        aliases["r-benezra@ski.mskcc.org"] = "benezrar"
-        aliases["r-kolesnick@ski.mskcc.org"] = "rkolesnick"
-        aliases["s-keeney@ski.mskcc.org"] = "keeneys"
-        aliases["s-shuman@ski.mskcc.org"] = "sshuman"
-        aliases["w-mark@ski.mskcc.org"] = "markw"
-
-        return aliases
+aliases = {
+    "a-haimovitz-friedman@ski.mskcc.org": "haimovia",
+    "a-zelenetz@mskcc.org": "zeleneta",
+    "a-zelenetz@ski.mskcc.org": "zeleneta",
+    "d-scheinberg@ski.mskcc.org": "scheinbd",
+    "f-giancotti@ski.mskcc.org": "giancotf",
+    "j-massague@ski.mskcc.org": "massaguej",
+    "k-anderson@ski.mskcc.org": "kanderson",
+    "m-baylies@ski.mskcc.org": "bayliesm",
+    "m-jasin@ski.mskcc.org": "jasinm",
+    "m-mcdevitt@ski.mskcc.org": "m-mcdevitt",
+    "m-moore@ski.mskcc.org": "moorem",
+    "m-ptashne@ski.mskcc.org": "ptashne",
+    "m-sadelain@ski.mskcc.org": "sadelaim",
+    "m-van-den-brink@ski.mskcc.org": "vandenbm",
+    "p-tempst@ski.mskcc.org": "tempstp",
+    "r-benezra@ski.mskcc.org": "benezrar",
+    "r-kolesnick@ski.mskcc.org": "rkolesnick",
+    "s-keeney@ski.mskcc.org": "keeneys",
+    "s-shuman@ski.mskcc.org": "sshuman",
+    "w-mark@ski.mskcc.org": "markw"
+}
